extract getRecordsFromEvent
